Tests for layered_test_check.py
