+++ edx_lint/pylint/layered_test_check.py
@@ -25,7 +25,7 @@ def is_test_case_class(node):
 
     dunder_test = node.locals.get("__test__")
     if dunder_test:
-        if isinstance(dunder_test[0], astroid.AssName):
+        if isinstance(dunder_test[0], astroid.AssignName):
             value = list(dunder_test[0].assigned_stmts())
             if len(value) == 1 and isinstance(value[0], astroid.Const):
                 return bool(value[0].value)
