--- test/messages/func_layered_tests.txt
@@ -1,2 +1,0 @@
-E: 14:DerivedTestCase: test class DerivedTestCase inherits tests from TestCase
-E: 43:TestsWithHelpers2: test class TestsWithHelpers2 inherits tests from TestHelpers2
