+++ tox.ini
@@ -41,3 +41,6 @@ filterwarnings =
     ;   .tox/py27/lib/python2.7/site-packages/_pytest/compat.py:329:
     ;       RemovedInPytest4Warning: usage of Session.Class is deprecated, please use pytest.Class instead
     ignore::pytest.PytestDeprecationWarning
+    ;   .tox/py35/lib/python3.5/site-packages/astroid/builder.py:231:
+    ;       PendingDeprecationWarning: generator 'wrapped' raised StopIteration
+    ignore::PendingDeprecationWarning:astroid
